Added a new drop-down to allow for filtering of animals by enclosure.
once a choice is selecteed, the page is reloaded and a POST request is sent. This post request is caught by the first block of pHp logic, and queries the appropriate group of animals to be displayed.

also added the animal grid logic so each animal gets it's own div with an image and a name. (image not added yet).